fix: 댓글 수정 시 댓글 이미지가 없는 경우 나오던 에러 해결

+++ client/src/main/java/com/fc8/service/impl/PostServiceImpl.java
@@ -222,9 +222,12 @@ public class PostServiceImpl implements PostService {
         // 3. 댓글 수정
         postComment.modify(command.getContent());
 
-        // 4. 기존 이미지 삭제 및 변경
-        postCommentImage.delete();
+        // 4. 기존 이미지가 있는 경우 기존 이미지 삭제 및 변경
+        if (postCommentImage != null) {
+            postCommentImage.delete();
+        }
 
+        // 5. 새로운 이미지가 있는 경우 이미지 저장
         Optional.ofNullable(image)
             .filter(img -> !img.isEmpty())
             .ifPresent(img -> {
